git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@13756 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/gamedata.cpp
@@ -2593,7 +2593,7 @@ if (nChecks) {
 		return (tRobotInfo*) GameDataError ("robot info", "buffer", pszFile, nLine);
 	if ((nChecks & GAMEDATA_CHECK_UNDERFLOW) && (nId < 0))
 		return (tRobotInfo*) GameDataError ("robot info", "underflow", pszFile, nLine);
-	if ((nChecks & GAMEDATA_CHECK_OVERFLOW) && (nId >= a.Length ()))
+	if ((nChecks & GAMEDATA_CHECK_OVERFLOW) && ((uint32_t) nId >= a.Length ()))
 		return (tRobotInfo*) GameDataError ("robot info", "overflow", pszFile, nLine);
 	}
 return a + nId;
